feat: Upload the thumbnail to pinata
Upload both - the movie file as well as the thumbnail to ipfs. setLink should be called as soon as the movie file is uploaded, hence switched uploadFilesToPinata to a stream generator which yields the link as soon as uploading is finished. This stream is consumed inside the handleSubmit function for Uploads.tsx inside try and catch blocks.